requirements.txt
demucs @ git+https://github.com/facebookresearch/demucs.git@583db9df0213ba5f5b3491eca5c993e7629f1949#egg=demucs
setuptools>=61.0
pytaglib==3.0.0
ffmpeg-python==0.2.0
torch>=2.1.2
torchaudio>=2.1.2
click==8.1.7
Cython==3.0.10
